refactored format defects to defect service
