+++ src/services/DefectService.ts
@@ -1,14 +1,21 @@
-import { Service } from 'typedi';
+import { Inject, Service } from 'typedi';
 import moment from 'moment';
 import { ICustomDefect } from '../models/ICustomDefect';
-import { IVA_30 } from '../models/Enums';
+import { IVA_30, LOCATION_WELSH } from '../models/Enums';
 import { IFlatDefect } from '../models/IFlatDefect';
 import { IItem } from '../models/IItem';
 import { IDefectParent } from '../models/IDefectParent';
 import { IDefectChild } from '../models/IDefectChild';
+import { TranslationService } from './TranslationService';
 
 @Service()
 export class DefectService {
+  private readonly translationService: TranslationService;
+
+  constructor(@Inject() translationService: TranslationService) {
+    this.translationService = translationService;
+  }
+
   /**
    * Calculates the retest date for an IVA or MSVA test
    * @param testTypeStartTimestamp - the test start timestamp of the test
@@ -155,5 +162,75 @@ export class DefectService {
       (flatDefect: IFlatDefect) => flatDefect.forVehicleType!.includes(vehicleType),
     );
     return filteredWelshDefectsOnVehicleType[0];
+  }
+
+  /**
+   * Returns a formatted welsh string containing data about a given defect
+   * @param defect - the defect for which to generate the formatted welsh string
+   * @param vehicleType - the vehicle type from the test result
+   * @param flattenedDefects - the list of flattened defects
+   */
+  public formatDefectWelsh(
+    defect: any,
+    vehicleType: any,
+    flattenedDefects: IFlatDefect[],
+  ) {
+    const toUpperFirstLetter: any = (word: string) => word.charAt(0).toUpperCase() + word.slice(1);
+
+    const filteredFlatDefects: IFlatDefect[] = flattenedDefects.filter(
+      (x: IFlatDefect) => defect.deficiencyRef === x.ref,
+    );
+
+    const filteredFlatDefect: IFlatDefect | null = this.filterFlatDefects(
+      filteredFlatDefects,
+      vehicleType,
+    );
+
+    if (filteredFlatDefect !== null) {
+      let defectString = `${defect.deficiencyRef} ${filteredFlatDefect.itemDescriptionWelsh}`;
+
+      if (defect.deficiencyText) {
+        defectString += ` ${filteredFlatDefect.deficiencyTextWelsh}`;
+      }
+
+      if (defect.additionalInformation.location) {
+        Object.keys(defect.additionalInformation.location).forEach(
+          (location: string, index: number, array: string[]) => {
+            if (defect.additionalInformation.location[location]) {
+              const welshLocation = this.translationService.convertLocationWelsh(
+                defect.additionalInformation.location[location],
+              );
+
+              switch (location) {
+                case 'rowNumber':
+                  defectString += ` ${LOCATION_WELSH.ROW_NUMBER}: ${defect.additionalInformation.location.rowNumber}.`;
+                  break;
+                case 'seatNumber':
+                  defectString += ` ${LOCATION_WELSH.SEAT_NUMBER}: ${defect.additionalInformation.location.seatNumber}.`;
+                  break;
+                case 'axleNumber':
+                  defectString += ` ${LOCATION_WELSH.AXLE_NUMBER}: ${defect.additionalInformation.location.axleNumber}.`;
+                  break;
+                default:
+                  defectString += ` ${toUpperFirstLetter(welshLocation)}`;
+                  break;
+              }
+            }
+
+            if (index === array.length - 1) {
+              defectString += '.';
+            }
+          },
+        );
+      }
+
+      if (defect.additionalInformation.notes) {
+        defectString += ` ${defect.additionalInformation.notes}`;
+      }
+      console.log(`Welsh Defect String Generated: ${defectString}`);
+      return defectString;
+    }
+    console.log('ERROR: Unable to find a filtered defect');
+    return null;
   }
 }
