+++ tests/unit/DefectService.unitTest.ts
@@ -3,24 +3,19 @@ import 'reflect-metadata';
 import Container from 'typedi';
 import sinon from 'sinon';
 import { cloneDeep } from 'lodash';
-import { CertificatePayloadGenerator } from '../../src/services/CertificatePayloadGenerator';
 import mockTestResult from '../resources/test-result-with-defect.json';
 import defectsMock from '../resources/defects_mock.json';
 import flatDefectsMock from '../resources/flattened-defects.json';
 import { LambdaMockService } from '../models/LambdaMockService';
 import { LambdaService } from '../../src/services/LambdaService';
 import { DefectService } from '../../src/services/DefectService';
-import { CertificatePayloadGeneratorPassOrFail } from '../../src/services/CertificatePayloadGeneratorPassOrFail';
 
 context('Defects', () => {
   describe('welsh defect function', () => {
     Container.set(LambdaService, new LambdaMockService());
-    const certGenSvc = Container.get(CertificatePayloadGenerator);
     const sandbox = sinon.createSandbox();
 
     const defectService = Container.get(DefectService);
-
-    const passOrFailCertificatePayloadGenerator = Container.get(CertificatePayloadGeneratorPassOrFail);
 
     afterEach(() => {
       sandbox.restore();
@@ -31,7 +26,7 @@ context('Defects', () => {
         // get mock of defect or test result
         const testResultWithDefect = cloneDeep(mockTestResult);
         console.log(testResultWithDefect.testTypes[0].defects[0]);
-        const format = passOrFailCertificatePayloadGenerator.formatDefectWelsh(
+        const format = defectService.formatDefectWelsh(
           testResultWithDefect.testTypes[0].defects[0],
           'hgv',
           flatDefectsMock,
@@ -46,7 +41,7 @@ context('Defects', () => {
         // get mock of defect or test result
         const testResultWithDefect = cloneDeep(mockTestResult);
         console.log(testResultWithDefect.testTypes[0].defects[0]);
-        const format = passOrFailCertificatePayloadGenerator.formatDefectWelsh(
+        const format = defectService.formatDefectWelsh(
           testResultWithDefect.testTypes[0].defects[0],
           'trl',
           flatDefectsMock,
@@ -61,7 +56,7 @@ context('Defects', () => {
         // get mock of defect or test result
         const testResultWithDefect = cloneDeep(mockTestResult);
         console.log(testResultWithDefect.testTypes[0].defects[0]);
-        const format = passOrFailCertificatePayloadGenerator.formatDefectWelsh(
+        const format = defectService.formatDefectWelsh(
           testResultWithDefect.testTypes[0].defects[0],
           'psv',
           flatDefectsMock,
@@ -79,7 +74,7 @@ context('Defects', () => {
         Object.assign(testResultWithDefect.testTypes[0].defects[0].additionalInformation.location, { seatNumber: 2 });
         Object.assign(testResultWithDefect.testTypes[0].defects[0].additionalInformation.location, { axleNumber: 3 });
         console.log(testResultWithDefect.testTypes[0].defects[0]);
-        const format = passOrFailCertificatePayloadGenerator.formatDefectWelsh(
+        const format = defectService.formatDefectWelsh(
           testResultWithDefect.testTypes[0].defects[0],
           'hgv',
           flatDefectsMock,
@@ -97,7 +92,7 @@ context('Defects', () => {
         // get mock of defect or test result
         const testResultWithDefect = cloneDeep(mockTestResult);
         console.log(testResultWithDefect.testTypes[0].defects[0]);
-        const format = passOrFailCertificatePayloadGenerator.formatDefectWelsh(
+        const format = defectService.formatDefectWelsh(
           testResultWithDefect.testTypes[0].defects[0],
           'hgv',
           [],
